% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{add_linters}
\alias{add_linters}
\title{Add linters}
\usage{
add_linters(config, tag, linters, overwrite = FALSE)
}
\arguments{
\item{config}{A config object to add linters to.}

\item{tag}{The name of the tag to which to add linters.}

\item{linters}{New linters to add.}

\item{overwrite}{A logical value indicating whether existing linters should
be overwritten.}
}
\value{
\code{NULL}.
}
\description{
Add linters
}
\keyword{internal}
