% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{capture_nolints}
\alias{capture_nolints}
\title{Capture \verb{# nolint} in strings}
\usage{
capture_nolints(x)
}
\arguments{
\item{x}{\code{character} vector, with strings possibly containing an
declaration.}
}
\value{
A \code{list} of length equal to that of parameter \code{x}, giving itemized
lint names in cases of matches, \code{TRUE} if a blanket \verb{# nolint} declaration
is provided and \code{FALSE} if there is no matched \verb{# nolint} declaration at
all.
}
\description{
Capture \verb{# nolint} in strings
}
