% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linter.R
\name{check_linter}
\alias{check_linter}
\title{Apply 'roxygen2' linters}
\usage{
check_linter(linters, tag, ...)
}
\arguments{
\item{linters}{A linters specification, either a \code{function} or
\code{character} regular expression that should be matched, or a \code{list}
of either. See details for more information.}

\item{tag}{A \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}.}
}
\value{
\code{TRUE}, invisibly. However, this function is primarily used for its
side-effect of raising alerts during documentation.
}
\description{
Provided a list of linters for a given tag, iterate over linters to raise
alerts during the documentation process.
}
