% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{config_tag_accepts_linters}
\alias{config_tag_accepts_linters}
\title{Check whether a given tag accepts new linters}
\usage{
config_tag_accepts_linters(config, tag)
}
\arguments{
\item{config}{A config object to add linters to.}

\item{tag}{The name of the tag to which to add linters.}
}
\description{
Linters can be configured to disallow new linter additions by providing a
trailing \code{NULL} value to a list, or setting the value itself to \code{NULL}.
}
\keyword{internal}
