% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{find_lint_generic}
\alias{find_lint_generic}
\title{Find the last called lint function}
\usage{
find_lint_generic(skip = 1)
}
\value{
\code{function} the most recent function on the stack of type \code{lint}.
}
\description{
Find the last called lint function
}
\keyword{internal}
