% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_alerts.R
\name{format_tag_prefix}
\alias{format_tag_prefix}
\title{Format a tag for alert context}
\usage{
format_tag_prefix(x)
}
\arguments{
\item{x}{A \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}.}
}
\value{
A formatted string
}
\description{
Format a tag for alert context
}
\keyword{internal}
