% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_alerts.R
\name{into_roxy_alert}
\alias{into_roxy_alert}
\title{Capture output and re-emit as a cli alert}
\usage{
into_roxy_alert(tag, expr)
}
\arguments{
\item{tag}{A \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}} to use as context for the alert.}

\item{expr}{An expression to evaluate.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
Capture output and re-emit as a cli alert
}
\keyword{internal}
