% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{is_lint_disabled}
\alias{is_lint_disabled}
\title{Check whether the current lint is disabled}
\usage{
is_lint_disabled(x, lint_name = attr(fn, "name"), fn = find_lint_generic())
}
\arguments{
\item{lint_name}{\code{character} name of the lint. Used if only specific lints
are disabled for the provided tag.}

\item{fn}{\code{function} to use to test for lint disabling. The
\code{"name"} attribute of this function will be used to infer the lint
name. Defaults to the most recent \code{lint} function evaluated on the call
stack.}
}
\value{
\code{logical} flag indicating whether this lint has been disabled for
the given tag.
}
\description{
Scans raw text of tag for \verb{`r # nolint` or `r # nolint: <type>.`}
}
\keyword{internal}
