% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\name{lint}
\alias{lint}
\title{Define a lint generic}
\usage{
lint(f, ...)
}
\description{
Attach optional attributes to a linting function. Most commonly used for
providing attributes:
}
\details{
\itemize{
\item \code{name}: A string used for disabling via \verb{# nolint: <name>.}
}
}
\keyword{internal}
