% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linters.R
\docType{data}
\name{linters}
\alias{linters}
\alias{lint_sentence_case}
\alias{tidy_title}
\alias{tidy_param}
\alias{tidy}
\title{Linters}
\format{
An object of class \code{list} of length 2.
}
\usage{
lint_sentence_case(x, ...)

tidy_title(x, ...)

tidy_param(x, name, description, ...)

tidy
}
\arguments{
\item{x}{The \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}} that prompted the lint.}

\item{...}{Additional arguments unused.}

\item{name, description}{Used for \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}-specific linters.}
}
\description{
Preconfigured linters, either as a collective list of linters or
individually. "tidy" linters implement guidelines from the tidyverse style
guide.
}
\details{
Refer to the individual \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}} for the respective tag for
argument details.
}
\section{Functions}{
\itemize{
\item \code{lint_sentence_case()}: Lint sentence case (uses \verb{$raw} for \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}s)

\item \code{tidy_title()}: Tidy 'Sentence case' titles

\item \code{tidy_param()}: Tidy 'Sentence case' \verb{@param} definitions

\item \code{tidy}: A list of all tidyverse style guide inspired linters

}}
\keyword{datasets}
