% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rd_text}
\alias{rd_text}
\title{Parse a \code{\link{roxygen2}} tag into its visible text}
\usage{
rd_text(
  rd,
  prune = c("\\\\code", "\\\\preformatted", "\\\\file", "\\\\email", "\\\\url",
    "\\\\cite", "\\\\acronym")
)
}
\arguments{
\item{tag}{A \code{\link{roxygen2}} tag. Expected to have an implementation of
\code{\link[roxygen2:roxy_tag_rd]{roxygen2::roxy_tag_rd}} and its result should implement \code{\link{format}}.}
}
\value{
A \code{character} string of visible text.
}
\description{
Optionally prunes specific \code{Rd} tags whose rendered content should not
affect linting.
}
\keyword{internal}
