% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register.R
\name{register_linters}
\alias{register_linters}
\title{Register linters for new tags}
\usage{
register_linters(..., linters = list(...), .overwrite = FALSE)
}
\arguments{
\item{...}{Linters to register, with parameter names used to identify the
associated tag.}

\item{linters}{Optionally, provide a named list of linters directly.}

\item{.overwrite}{Whether existing linters should be overwritten. Intended to
be used sparingly. For example, this could be used to implement an entire new
style guide without inheriting defaults.}
}
\value{
\code{TRUE}, invisibly.
}
\description{
A registration mechanism for other packages to provide linters, either for
custom tags or to implement their own linting styles.
}
