% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxy_tag_rd.R
\name{roxy_tag_rd}
\alias{roxy_tag_rd}
\title{Modified linting roxy tag evaluation}
\usage{
roxy_tag_rd(x, base_path, env)
}
\arguments{
\item{x}{The tag}

\item{base_path}{Path to package root directory.}

\item{env}{Environment in which to evaluate code (if needed)}
}
\value{
A processed \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}
}
\description{
Intercept \code{roxygen2} to perform linting.
}
\keyword{internal}
