% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet.R
\name{roxylint}
\alias{roxylint}
\title{Roclet for 'roxygen2' style linting}
\usage{
roxylint()
}
\value{
A \code{roxylint} \code{\link[roxygen2:roclet]{roxygen2::roclet()}}.
}
\description{
Roclets used to embed linters during documentation. To use, add the roclet in
your \code{DESCRIPTION} file.
}
\details{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Config/Needs/documentation: roxylint
Roxygen:
  list(
    markdown = TRUE,
    roclets = c("namespace", "rd", "roxylint::roxylint")
  )
}\if{html}{\out{</div>}}
}
