% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scrub_markup}
\alias{scrub_markup}
\title{Scrubs a string of texinfo and markdown markup}
\usage{
scrub_markup(x)
}
\description{
Superceded by \code{tag_text} but kept around in case the alternatives ever
break down for some edge case.
}
\keyword{internal}
