% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linter.R, R/roclet.R
\name{check_linter.default}
\alias{check_linter.default}
\alias{check_linter.list}
\alias{check_linter.function}
\alias{check_linter.character}
\alias{roxylint}
\alias{roclet_process.roclet_roxylint}
\alias{roclet_output.roclet_roxylint}
\title{Roclet for 'roxygen2' style linting}
\usage{
\method{check_linter}{default}(linters, tag, ...)

\method{check_linter}{list}(linters, tag, ...)

\method{check_linter}{`function`}(linters, tag, ...)

\method{check_linter}{character}(linters, tag, message = NULL)

roxylint()

roclet_process.roclet_roxylint(x, blocks, env, base_path)

roclet_output.roclet_roxylint(...)
}
\value{
A \code{roxylint} \code{\link[roxygen2:roclet]{roxygen2::roclet()}}.
}
\description{
Roclet for 'roxygen2' style linting
}
\section{Functions}{
\itemize{
\item \code{check_linter(default)}: By default, no linting is performed

\item \code{check_linter(list)}: A \code{list} of \code{function}s or \code{character} regular expressions.

If a \code{character} value is named, the name is used as the message for a alert
that is raised when the expression does not match.

\item \code{check_linter(`function`)}: A \code{function} to evaluate for the given tag

\code{function}'s are evaluated with the following arguments:
\enumerate{
\item The \code{\link[roxygen2:roxy_tag]{roxygen2::roxy_tag()}}
\item The contents of the tag's \verb{$val}, as named arguments
}

Because the number of arguments might not be readily apparent, any function
should accept a trailing \code{...} argument.

Provided \code{function}s may print lint output, or signal lint output with
messages or warnings. A \code{\link[cli:cli_alert]{cli::cli_alert()}} will reflect the severity of the
function used to emit the output.

\item \code{check_linter(character)}: A \code{character} regular expressions.

If a \code{character} value is found, its value is assumed to be a regular
expression which must match a given tag's \code{raw} content (the text as it
appears in the \code{roxygen2} header).

\item \code{roclet_process.roclet_roxylint()}: For each tag scanned by 'roxygen2', apply any linters

\item \code{roclet_output.roclet_roxylint()}: Unused

}}
